feat: create consent component
